% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_dm.R
\name{acs_attr}
\alias{acs_attr}
\title{Add default attributes for ACS tables}
\usage{
acs_attr(acs_year, acs_state_code, acs_geography, acs_county, lt6 = "002", pth)
}
\description{
Add default attributes for ACS tables
}
