% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_child_care_db.R
\name{acs_tables}
\alias{acs_tables}
\title{Add default parameters for ACS tables}
\usage{
acs_tables(acs_year, acs_state_code, acs_geography, acs_county, raw_pth)
}
\description{
Add default parameters for ACS tables
}
