% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_joins.R
\name{anchor_tract_sub}
\alias{anchor_tract_sub}
\title{Filter to selected anchor_tract}
\usage{
anchor_tract_sub(df_anchor_tract, input_mkt = NULL)
}
\arguments{
\item{input_mkt}{data.frame. Data frame with subset of tracts to map}

\item{df_tracts_xwalks}{data.frame. Crosswalk between the anchor and surround
tracts}
}
\description{
Filter to selected anchor_tract
}
