% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_dm.R
\name{assign_acf_class}
\alias{assign_acf_class}
\title{Assigns a class to ACF data}
\usage{
assign_acf_class(pth, acf_qtr_years, max_child_age)
}
\arguments{
\item{pth}{String. Path to the data}

\item{acf_qtr_years}{vector. Default is 'Q2-2019'.}
}
\description{
Assigns a class to all the incoming files which will perform different
data management steps according to which file type it is
}
