% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{attr.ccl}
\alias{attr.ccl}
\title{Frontline attributes}
\usage{
attr.ccl(
  pth,
  state_code,
  df_twc,
  url = "https://data.texas.gov/api/views/bc5r-88dy/rows.csv?accessType=DOWNLOAD",
  ext = "csv",
  input_columns = list(OPERATION_NUMBER = "character", OPERATION_NAME = "character",
    OPERATION_TYPE = "character", Location_address_geo = "character", COUNTY =
    "character", TOTAL_CAPACITY = "numeric", LICENSED_TO_SERVE_AGES = "character",
    PROGRAMS_PROVIDED = "character", ACCEPTS_CHILD_CARE_SUBSIDIES = "character",
    email_address = "character", PHONE_NUMBER = "character"),
  naeyc_pth1 = "BP4K - NAEYC List - 5.26.21.xlsx",
  naeyc_pth2 = "NAEYC Providers - NAEYC - 5.26.21.xlsx",
  name = "HHSC_CCL",
 
    bb_url = "https://gist.githubusercontent.com/a8dx/2340f9527af64f8ef8439366de981168/raw/81d876daea10eab5c2675811c39bcd18a79a9212/US_State_Bounding_Boxes.csv",
  geocode = list(url = "http://www.mapquestapi.com", path = "/geocoding/v1/batch",
    limit = 100, qualityCode = "A1|A3|A4"),
  reverse_geocode = list(url = "https://geo.fcc.gov", path = "/api/census/block/find")
)
}
\arguments{
\item{pth}{string. The path to write the raw to.}

\item{state_code}{string. The state fips code}

\item{df_twc}{data.frame. TWC data}

\item{url}{string. The url to the data.}

\item{input_columns}{list. The columns to expect and keep from the data.}

\item{naeyc_pth1}{string.}

\item{naeyc_pth2}{string.}

\item{name}{string. The name to write the raw data to.}

\item{bb_url}{string. Bounding box url.}
}
\value{
object
}
\description{
Frontline attributes
}
