% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{attr.frontline}
\alias{attr.frontline}
\title{Frontline attributes}
\usage{
attr.frontline(
  pth,
  input_columns = list(`OP Number` = "character", `Infant Capacity` = "numeric",
    `Toddler Capacity` = "numeric", `Pre- K Capacity` = "numeric", `School Aged Capacity`
    = "numeric", Infant_enrollment = "numeric", Toddler_enrollment = "numeric",
    Preschool_enrollment = "numeric", SchoolAge_enrollment = "numeric",
    last_modified_at_A = "character", Date = "character", export_date = "Date"),
  name = "frontline/export_translation_Daily_Vacancy_2021-09-07_05_12_48.csv",
  max_date = "08042021"
)
}
\arguments{
\item{input_columns.}{The columns to expect and keep from the data.}

\item{name.}{The name of the data.}

\item{max_date.}{Two week period with the highest reponse rate. Default is 
08022021.}
}
\value{
object
}
\description{
Frontline attributes
}
