% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{attr.twc}
\alias{attr.twc}
\title{Frontline attributes}
\usage{
attr.twc(
  pth,
  name = "Weekly Closure Report 9.7.21.xlsx",
  sheet = "All Provider Level",
  na = "NA",
  data_name = "TWC",
  input_columns = list(`License Number` = "numeric", TWC = "character", `TRS Flag` =
    "character", `Number of Current Referrals` = "numeric", twc_date = "Date"),
  twc_var = list(var_name = "twc", values = c("Y", "N")),
  trs_flag_var = list(var_name = "trs_flag", values = c("Regular", "TRS 2 Star",
    "TRS 4 Star", "TRS 3 Star"))
)
}
\arguments{
\item{name}{string. Name of the data to use.}

\item{sheet}{string. Name of the sheet in the excel file to use.}

\item{na.}{string. Values to assign to NA when reading in data.}

\item{input_columns.}{The columns to expect and keep from the data.}
}
\value{
object
}
\description{
Frontline attributes
}
