% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkt_analysis.R
\name{avg_provider_mkt}
\alias{avg_provider_mkt}
\title{Find average number of providers per market}
\usage{
avg_provider_mkt(xwalk_tract_provider)
}
\description{
Find average number of providers per market
}
