% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subsidy_capacity.R
\name{calc.capacity_adjustment_sub}
\alias{calc.capacity_adjustment_sub}
\title{Create subsidy capacity estimate}
\usage{
calc.capacity_adjustment_sub(
  df_hhsc_ccl,
  df_frontline,
  df_supply_adjustment_03 = NULL,
  grouping_vars = NULL
)
}
\arguments{
\item{df_hhsc_ccl}{data.frame.}

\item{df_frontline}{data.frame.}

\item{grouping_vars}{string. The results to be grouped by. Default is NULL.}
}
\description{
Create subsidy capacity estimate
}
