% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tract_radius.R
\name{calc.distance_density_plot}
\alias{calc.distance_density_plot}
\title{Distance density plot}
\usage{
calc.distance_density_plot(df, county_list)
}
\arguments{
\item{df.}{data.frame. Default is DF_ACF.}

\item{county_list.}{A list of TX county codes of counties of interest to 
compare. E.g county_list = c(48439) for Tarrant County; 
county_list = c("48201", "48439") to compare Harris to Tarrant County}
}
\value{
plot
}
\description{
Distance density plot
}
