% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subsidy_capacity.R
\name{calc.subsidy_capacity}
\alias{calc.subsidy_capacity}
\title{Create subsidy capacity estimate}
\usage{
calc.subsidy_capacity(
  config,
  xwalk_tracts,
  adj_tracts,
  df_hhsc_ccl,
  df_acf,
  grouping_vars = NULL,
  qtrs = c("1", "2", "4")
)
}
\arguments{
\item{config}{list. Vector of county FIPS codes with names attributes}

\item{xwalk_tracts}{data.frame.}

\item{adj_tracts}{data.frame.}

\item{df_hhsc_ccl}{data.frame.}

\item{df_acf}{data.frame.}

\item{grouping_vars}{string. The results to be grouped by. Default is NULL.}

\item{qtrs}{vector of strings. Default is c("1","2","4").}

\item{tract_radius}{numeric. A number indicating the tract radius in miles.}
}
\description{
Create subsidy capacity estimate
}
