% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.no_na}
\alias{check.no_na}
\title{Checks there are no NAs}
\usage{
\method{check}{no_na}(
  df,
  var_attr,
  data_name,
  msg = "Missing values in {var_name} variable in {data_name} data"
)
}
\description{
Checks there are no NAs
}
