% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.values}
\alias{check.values}
\title{Checks the values match what is expected}
\usage{
\method{check}{values}(
  df,
  var_attr,
  data_name,
  msg = "Values other than {v_collapse} in {var_name} in {data_name} data"
)
}
\arguments{
\item{data_name}{string. Name of the data to include in message if assert 
fails.}

\item{var}{vector. List of values to check.}

\item{values}{vector. List of values to check against.}

\item{var_name}{string. Name of the variable to include in message if assert 
fails.}
}
\description{
Checks the values match what is expected
}
