% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_dm.R
\name{check_census_key}
\alias{check_census_key}
\title{Checks to make sure the census API key is set as an environment variable}
\usage{
check_census_key(env_var = "CENSUS_API_KEY")
}
\arguments{
\item{env_var.}{String. The environment variable to check for. Default is CENSUS_API_KEY}
}
\value{
TRUE if a census api key exists, error if it does not exist
}
\description{
Checks to make sure the census API key is set as an environment variable
}
\examples{
\dontrun{
key <- "X"
tidycensus::census_api_key(key)
check_census_key()
}
}
