% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_dm.R
\name{check_county_bounds}
\alias{check_county_bounds}
\title{Check county bounds}
\usage{
check_county_bounds(df, county_bb)
}
\value{
data.frame
}
\description{
Checks that non-missing lat and longitudes are within their designated county boundaries
and if they are not it assign an NA
}
