% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_dm.R
\name{check_tx_bounds}
\alias{check_tx_bounds}
\title{Check Texas Bounds}
\usage{
check_tx_bounds(x)
}
\value{
data.frame
}
\description{
Checks that non-missing lat and longitudes are within the Texas
state boundaries and if they are not it assign an NA
}
