% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cols.R
\name{check_type}
\alias{check_type}
\title{Tests the vector or column type}
\usage{
check_type(v, msg = NULL)
}
\arguments{
\item{v.}{Vector. A vector to test.}

\item{msg.}{String. A message to return if the check fails.}
}
\description{
Tests the vector or column type
}
