% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.accepts_child_care_subsidies}
\alias{col.accepts_child_care_subsidies}
\title{Data management steps to clean the accepts child care subsidies column}
\usage{
col.accepts_child_care_subsidies(x)
}
\arguments{
\item{x}{object.}
}
\value{
object
}
\description{
Data management steps to clean the accepts child care subsidies column
}
