% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.assign_deserts}
\alias{col.assign_deserts}
\title{Assign deserts}
\usage{
col.assign_deserts(df, df_twc, naeyc_pth1, naeyc_pth2)
}
\description{
Assign deserts based on provider types. Note Head Start is 
included as a subsidy provider, a TRS and TRS 4 star provider.
}
