% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{col.availability}
\alias{col.availability}
\title{Data management for availability data}
\usage{
col.availability(df)
}
\value{
data.frame
}
\description{
Manage availability: sum 0-3 availability and total availability
}
