% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{col.confirm_sub}
\alias{col.confirm_sub}
\title{Confirm subsidy provider if provider is TRS status}
\usage{
col.confirm_sub(x)
}
\description{
Function replaces any provider who has a TRS status to also take
subsidy. This step is done because of data quality issue from TWC. In order 
to be a TRS provider, the provider must take the subsidy, so we know this is t
true.
}
