% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{col.county_twc}
\alias{col.county_twc}
\title{Data management steps for the county column}
\usage{
col.county_twc(df, state_fips)
}
\value{
data.frame
}
\description{
Data management steps for the county column
}
