% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{col.date}
\alias{col.date}
\title{Data management for date column}
\usage{
col.date(df)
}
\value{
data.frame
}
\description{
Manage date: filter date to be less than export date so that future dates for which 
providers have not yet entered data will be excluded.
}
