% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.dm_naeyc}
\alias{col.dm_naeyc}
\title{Data management steps to add the naeyc column}
\usage{
col.dm_naeyc(x)
}
\arguments{
\item{x}{object.}
}
\value{
object
}
\description{
Assigns NAEYC status to providers from different sources 
provided by Tarrant County. Note: we are not sure how often these will be 
updated and the code may have to change since there likely is not a 
consistent file format for these files.
}
