% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{col.enrollment}
\alias{col.enrollment}
\title{Data management for enrollment data}
\usage{
col.enrollment(x)
}
\value{
object
}
\description{
Manage enrollment: sum 0-3 enrollment and total enrollment
}
