% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.licensed_to_serve_ages}
\alias{col.licensed_to_serve_ages}
\title{Data management steps to clean the licensed to serve ages column}
\usage{
col.licensed_to_serve_ages(df)
}
\arguments{
\item{df}{data.frame. The dataframe}
}
\value{
data.frame
}
\description{
Data management steps to clean the licensed to serve ages column
}
