% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.location_address_geo}
\alias{col.location_address_geo}
\title{Data management steps to clean the operation type column}
\usage{
col.location_address_geo(x)
}
\arguments{
\item{x}{object.}
}
\value{
object
}
\description{
Data management steps to clean the operation type column
}
