% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.operation_number}
\alias{col.operation_number}
\title{Data management steps for the operation number column}
\usage{
col.operation_number(x)
}
\value{
data.frame
}
\description{
Data management steps for the operation number column
}
