% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.operation_type}
\alias{col.operation_type}
\title{Data management steps to clean the operation_type column}
\usage{
col.operation_type(x)
}
\arguments{
\item{x}{object}
}
\value{
object
}
\description{
Data management steps to clean the operation_type column
}
