% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.programs_provided}
\alias{col.programs_provided}
\title{Data management steps to clean the programs provided column}
\usage{
col.programs_provided(x)
}
\arguments{
\item{x}{object.}
}
\value{
object
}
\description{
Data management steps to clean the programs provided column
}
