% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{col.provider_type}
\alias{col.provider_type}
\title{Data management steps to clean the provider_type column}
\usage{
col.provider_type(df)
}
\value{
data.frame
}
\description{
Data management steps to clean the provider_type column
}
