% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.quality}
\alias{col.quality}
\title{Data management quality of the provider}
\usage{
col.quality(x)
}
\description{
Assigns a quality for the provider. Note Head Start is 
included as a subsidy provider, a TRS and TRS 4 star provider.
}
