% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.resolve_twc}
\alias{col.resolve_twc}
\title{Resolve TWC}
\usage{
col.resolve_twc(x)
}
\description{
Resolve differences between TWC and CCL data because some of the
providers in CCL are not in TWC. We assume CCL providers which are not in TWC
data are not TRS providers.
}
