% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.total_capacity}
\alias{col.total_capacity}
\title{Data management steps to clean the total capacity column}
\usage{
col.total_capacity(df)
}
\arguments{
\item{df}{data.frame. The dataframe}
}
\value{
data.frame
}
\description{
Data management steps to clean the total capacity column
}
