% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{col.total_capacity_twc}
\alias{col.total_capacity_twc}
\title{Data management steps to clean the capacity column}
\usage{
col.total_capacity_twc(df)
}
\value{
data.frame
}
\description{
Data management steps to clean the capacity column
}
