% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{col.trs_provider}
\alias{col.trs_provider}
\title{Data management for trs_provider column}
\usage{
col.trs_provider(df)
}
\value{
data.frame
}
\description{
Manage trs status: binary status, convert star level to num
}
