% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_dm.R
\name{county_bounding_box}
\alias{county_bounding_box}
\title{Pulls down bounding box parameters for each county in Texas}
\usage{
county_bounding_box(
 
    url = "https://raw.githubusercontent.com/stucka/us-county-bounding-boxes/master/bounding.csv",
  state_fips
)
}
\description{
Pulls down bounding box parameters for each county in Texas
}
