% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_folder_str}
\alias{create_folder_str}
\title{Create data folder structure}
\usage{
create_folder_str(root, pths)
}
\arguments{
\item{root}{string. The root directory}

\item{pth}{list. A list of paths to create in the root directory}
}
\description{
Create data folder structure
}
