% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkt_analysis.R
\name{create_pct_dsrt_prvdr}
\alias{create_pct_dsrt_prvdr}
\title{Percent of deserts each provider serves in}
\usage{
create_pct_dsrt_prvdr(mkt_ratio, df_supply, xwalk_tracts)
}
\description{
Percent of deserts each provider serves in
}
