% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_joins.R
\name{desert_sub}
\alias{desert_sub}
\title{Filters the desert choices to a single desert}
\usage{
desert_sub(df_desert, input_dsrt)
}
\arguments{
\item{df_desert}{data.frame. A data frame with all the desert choices.}

\item{input_dsrt}{string. The desert to display.}
}
\description{
Filters the desert choices to a single desert
}
