% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tract_radius.R
\name{distance_decile_table}
\alias{distance_decile_table}
\title{Distance decile table}
\usage{
distance_decile_table(df = DF_ACF, county_list)
}
\arguments{
\item{df.}{data.frame. Default is DF_ACF.}

\item{county_list.}{A list of TX county codes of counties of interest to 
compare. E.g county_list = c(48439) for Tarrant County; 
county_list = c("48201", "48439") to compare Harris to Tarrant County}
}
\value{
table
}
\description{
Distance decile table
}
