% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_dm.R
\name{dm.acf}
\alias{dm.acf}
\title{Data management ACF}
\usage{
\method{dm}{acf}(raw_pth, acf_qtr_years, ...)
}
\arguments{
\item{pth}{String. Path to the data}
}
\value{
data.frame
}
\description{
Data are located: 
https://www.twc.texas.gov/programs/childcare#dataAndReports
}
