% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tract_radius.R
\name{dm.acf_dist}
\alias{dm.acf_dist}
\title{ACF data management to get zip codes for providers}
\usage{
\method{dm}{acf_dist}(df, max_child_age = 4)
}
\arguments{
\item{df}{data.frame. The dataframe of ACF data.}

\item{max_child_age.}{Integer. Age to filter the child age to be less than or equal.}
}
\value{
data.frame
}
\description{
Merge home and child care zip codes and calculate distance
}
