% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subsidy_capacity.R
\name{dm.agg_kids_prvdr}
\alias{dm.agg_kids_prvdr}
\title{Aggregate number of kids per provider}
\usage{
\method{dm}{agg_kids_prvdr}(df_acf, yr, qtrs)
}
\arguments{
\item{acf}{data.frame. The cleaned acf dataframe.}
}
\value{
Summarized data with the max, median, and minimum number of kids per
provider
}
\description{
Aggregate number of kids per provider
}
