% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_dm.R
\name{dm.B23008}
\alias{dm.B23008}
\title{Demand using table B23008}
\usage{
\method{dm}{B23008}(x)
}
\arguments{
\item{x}{object. The data object}
}
\description{
Creates variables: Estimated number of children under 5,
estimated number of children less than 6 with working parents, and
estimated number of children less than 5 with working parents
}
