% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_dm.R
\name{dm.drop_poor_quality}
\alias{dm.drop_poor_quality}
\title{Drops poor quality geocodes}
\usage{
\method{dm}{drop_poor_quality}(df, qualityCode = "A1|A3|A4")
}
\description{
Use the geocodeQualityCode value returned to determine the quality of the geocode. https://developer.mapquest.com/documentation/geocoding-api/quality-codes/.
}
