% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{dm.frontline}
\alias{dm.frontline}
\title{Fronline provider data management steps}
\usage{
\method{dm}{frontline}(x)
}
\description{
Data management steps, including, specifically enrollment and 
attendance, summation of seats
}
