% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_dm.R
\name{dm.geocode_address}
\alias{dm.geocode_address}
\title{Geocode addresses}
\usage{
\method{dm}{geocode_address}(x)
}
\arguments{
\item{addresses}{vector. The list of addresses to geocode.}

\item{key}{string. The api key registered with your personal Mapquest account.}
}
\description{
Geocodes addresses using the Mapquest API
}
