% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{dm.hhsc_ccl}
\alias{dm.hhsc_ccl}
\title{HHSC CCL data management}
\usage{
\method{dm}{hhsc_ccl}(x)
}
\value{
data.frame
}
\description{
Clean CCL download data, convert key variables to binary and 
select variables
}
