% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subsidy_capacity.R
\name{dm.mkt_enrollment_ratio}
\alias{dm.mkt_enrollment_ratio}
\title{Calculate enrollment ratios by market}
\usage{
\method{dm}{mkt_enrollment_ratio}(df_acf, yr, qtrs, df_hhsc_ccl, xwalk_tract_prvdr)
}
\description{
Calculate enrollment ratios by market
}
