% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_dm.R
\name{dm.mkt_subsidy}
\alias{dm.mkt_subsidy}
<<<<<<< HEAD
\title{Create market subsidy}
\usage{
dm.mkt_subsidy(acf, tracts_xwalk, cpp)
}
\description{
Create market subsidy
=======
\title{}
\usage{
\method{dm}{mkt_subsidy}(tracts, tract_provider_xwalk)
}
\description{

>>>>>>> main
}
