% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_dm.R
\name{dm.reverse_geocode}
\alias{dm.reverse_geocode}
\title{Geocode addresses}
\usage{
\method{dm}{reverse_geocode}(
  latLng,
  version = "v1",
  url = "http://www.mapquestapi.com/geocoding/{version}/reverse?key={key}"
)
}
\arguments{
\item{latLng}{list. The list of addresses to geocode.}

\item{key}{string. The api key registered with your personal Mapquest account.}
}
\description{
Geocodes addresses using the Mapquest API
}
\examples{
 latLng <- list(lat = 30.333472, lng = -81.470448)
 key <- "XXXXX"
 address <- dm.reverse_geocode(latLng = latLng, key = key)
}
