% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts_dm.R
\name{dm.tracts_xwalk}
\alias{dm.tracts_xwalk}
\title{Data management steps for the tracts crosswalk}
\usage{
\method{dm}{tracts_xwalk}(x, side = "right", width = 11, pad = "0")
}
\arguments{
\item{df}{data.frame. The tracts data to process}

\item{pth}{string. The name of the data to read in.}

\item{name}{string. The name of the data to write out.}
}
\description{
Creates a crosswalk between anchor tracts and surround tracts within x miles of each anchor tract
}
