% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{dm.twc}
\alias{dm.twc}
\title{Data management for TWC provider data}
\usage{
\method{dm}{twc}(
  df,
  input_columns = list(`License Number` = "numeric", TWC = "character", `TRS Flag` =
    "character", `Number of Current Referrals` = "numeric", Status = "character",
    `Reported Capacity` = "numeric", `Provider Type` = "character", County = "character",
    `Provider Name` = "character", `Board Name` = "character", twc_date = "Date"),
  state_fips
)
}
\description{
Manage TWC data, specifically TRS status and level, subsidy status, and
number of subsidy kids served
}
