% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwnld.acf}
\alias{dwnld.acf}
\title{Download ACF data}
\usage{
dwnld.acf(
  raw_pth,
  endpoint = "https://www.twc.texas.gov{path}",
  sheet = "ChildrenParentsSettings",
  ...
)
}
\description{
Data are located: 
https://www.twc.texas.gov/programs/childcare#dataAndReports
}
