% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_dm.R
\name{dwnld.acs}
\alias{dwnld.acs}
\title{Download ACS}
\usage{
dwnld.acs(x, ...)
}
\arguments{
\item{tbls}{list. List of census tables with attributes to download}

\item{pth.}{Path to save the raw data.}
}
\description{
Passes in a list of parameters to download the ACS census
data using functions from the tidycensus package
}
\details{
To find a list of parameters to pass see documentation for 
tidycensus::get_acs
}
\examples{
\dontrun{
x <- list(B23008 = list(year = 2019, state = 48, 
                               geography = "tract", county = 439))
dwnld.acs(x = x)
}
}
