% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{dwnld.frontline}
\alias{dwnld.frontline}
\title{Download Frontline provider data}
\usage{
dwnld.frontline(x)
}
\value{
object
}
\description{
Frontline provider data comes from the childcare.bowtiebi.com 
portal (currently, only Sadie has login)
}
