% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwnld.geo_county}
\alias{dwnld.geo_county}
\title{Get tract by latitude and longitude}
\usage{
dwnld.geo_county(state_fips)
}
\description{
Downloads shape file for Texas (48) using the tigris package, 
which pulls the most recent shape from the United States Census Bureau.
}
