% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwnld.geo_zip}
\alias{dwnld.geo_zip}
\title{Get ZCTA (Zip Code) and Latitude/Longitude Coordinate Crosswalk}
\usage{
dwnld.geo_zip(state_fips)
}
\description{
Downloads the crosswalk between zip codes and latitude longitude
coordinates of the center for Texas
}
