% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwnld.harris_neighborhood}
\alias{dwnld.harris_neighborhood}
\title{Get the Harris County Neighborhood to census tract data}
\usage{
dwnld.harris_neighborhood(
  url = "https://www.datahouston.org/cta_crosswalk.txt",
  name = "HARRIS_NEIGHBORHOOD",
  ext = "csv",
  raw_pth
)
}
\value{
data.frame
}
\description{
Download the neighbordhood to census tract cross walk create by 
the Kinder Institute. 
https://www.arcgis.com/apps/MapSeries/index.html?appid=95320b06677c438d91027cb5feb241bf
}
