% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{dwnld.hhsc_ccl}
\alias{dwnld.hhsc_ccl}
\title{Get HHSC CCL data}
\usage{
dwnld.hhsc_ccl(name, pth, url, ...)
}
\arguments{
\item{name}{string. The name to write the raw data to.}

\item{pth}{string. The path to write the raw to.}
}
\description{
Returns the most recent HHSC CCL Daycare and Residential Operations 
Data. Link to data: https://data.texas.gov/resource/bc5r-88dy.csv
}
