% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prek.R
\name{dwnld.isd}
\alias{dwnld.isd}
\title{Download ISD characteristics}
\usage{
dwnld.isd(raw_pth, name = "Current_Schools2020to2021.csv")
}
\description{
https://schoolsdata2-93b5c-tea-texas.opendata.arcgis.com/datasets/TEA-Texas::current-schools2020to2021/about
}
