% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwnld.lu_county_code}
\alias{dwnld.lu_county_code}
\title{Get State FIPS and State Name Crosswalk}
\usage{
dwnld.lu_county_code(state_fips)
}
\description{
Downloads the crosswalk between county fips codes and county 
names for Texas
}
