% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prek.R
\name{dwnld.prek}
\alias{dwnld.prek}
\title{Download prek}
\usage{
dwnld.prek(
  raw_pth,
  name = "Enrollment Report_Statewide_Campuses_Grade_2020-2021.csv"
)
}
\description{
data downloaded from
https://rptsvr1.tea.texas.gov/adhocrpt/adste.html
}
\note{
School Year: 2020-2021
Report: Statewide Campus Totals
Grade, Ethnicity or Gender: Grade
Downloaded as CSV file
}
