% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prek.R
\name{dwnld.prek_3}
\alias{dwnld.prek_3}
\title{Download prek 3 year old data}
\usage{
dwnld.prek_3(raw_pth, name = "artf392256_21.csv")
}
\description{
data requested from TEA through Public Information Request
}
\note{
School Year: 2020-2021
Downloaded as CSV file
}
