% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prek.R
\name{dwnld.rating}
\alias{dwnld.rating}
\title{Download school ratings}
\usage{
dwnld.rating(raw_pth, name = "Multi-year Rating List 2020.xlsx")
}
\description{
data downloaded from
https://rptsvr1.tea.texas.gov/perfreport/account/index.html
}
\note{
Schools were not rated in 2020 - use rating from 2019
}
