% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi.R
\name{dwnld.svi}
\alias{dwnld.svi}
\title{Download SVI}
\usage{
dwnld.svi(raw_pth, name = "Texas.csv")
}
\description{
data downloaded from
https://www.atsdr.cdc.gov/placeandhealth/svi/data_documentation_download.html
}
\note{
Year: 2018
Geography: Texas
Geography type: Census tracts
File type: CSV
}
