% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwnld.tarrant_neighborhood}
\alias{dwnld.tarrant_neighborhood}
\title{Get the Tarrant County Neighborhood data}
\usage{
dwnld.tarrant_neighborhood(
 
    url = "https://data.fortworthtexas.gov/api/views/ruhd-2sjc/rows.csv?accessType=DOWNLOAD",
  name = "TARRANT_NEIGHBORHOOD",
  ext = "csv",
  raw_pth
)
}
\value{
data.frame
}
\description{
Get the Tarrant County Neighborhood data
}
