% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts_dm.R
\name{dwnld.xwalk_tracts}
\alias{dwnld.xwalk_tracts}
\title{Get NBER Tract data}
\usage{
dwnld.xwalk_tracts(
  raw_pth,
  url = "http://data.nber.org/distance/2010/sf1/tract/sf12010tractdistance25miles.csv"
)
}
\arguments{
\item{url}{string. The URL to download the data from.}

\item{pth}{string. The raw path to write the data to.}
}
\description{
Link to data: http://data.nber.org/distance/2010/sf1/tract/sf12010tractdistance25miles.csv
}
\details{
25 mile radius is used rather then the 5 miles radius because the 5 mile radius was missing two counties in Harris County
}
