% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwnld.xwalk_zip_tract}
\alias{dwnld.xwalk_zip_tract}
\title{Get census tract and ZCTA (zip code) Crosswalk}
\usage{
dwnld.xwalk_zip_tract(
  pth = "https://www2.census.gov/geo/docs/maps-data/data/rel/zcta_tract_rel_10.txt",
  state_fips
)
}
\arguments{
\item{pth}{Link to data: 
https://www2.census.gov/geo/docs/maps-data/data/rel/zcta_tract_rel_10.txt}

\item{data_in_name}{string. The name to of the data to read in.}

\item{data_in_pth}{string. The path to read the data in from.}
}
\description{
Downloads the crosswalk between census tract codes and ZCTA 
for Texas.
}
