% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_codes.R
\name{dwnld.zip}
\alias{dwnld.zip}
\title{Download tract to zip xwalk}
\usage{
dwnld.zip(raw_pth, name = "TRACT_ZIP_062021.csv")
}
\description{
data downloaded from
https://www.huduser.gov/portal/datasets/usps_crosswalk.html
}
