% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{dwnld_df}
\alias{dwnld_df}
\title{Downloads data from a url and saves data}
\usage{
dwnld_df(url, pth, df_name = NULL, update = FALSE, unzip_fl = FALSE, ...)
}
\arguments{
\item{url}{string. The URL to get the data from}

\item{pth}{string. The path to write the data to}

\item{df_name}{string. The name to save the data to.}

\item{ext}{string. File extension.
will skip the first x (specified rows).}

\item{update.}{boolean. Indicates if the csv file should be updated.}

\item{unzip.}{logical. TRUE if the file needs to be unzipped.}
}
\description{
Downloads data from a url and saves data
}
