% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get.acf_data}
\alias{get.acf_data}
\title{Get ACF data}
\usage{
get.acf_data(
  data_in_name,
  data_in_pth,
  url = "https://www.twc.texas.gov/files/partners/{fl}",
  qtr = 1:4
)
}
\arguments{
\item{data_in_name}{string. The name to of the data to read in.}

\item{data_in_pth}{string. The path to read the data in from.}
}
\description{
Link to data: https://www.twc.texas.gov/programs/childcare#dataAndReports
}
