% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get.kinder_neighborhood_tract_xwalk}
\alias{get.kinder_neighborhood_tract_xwalk}
\title{Get the neighborhood to census tract data}
\usage{
get.kinder_neighborhood_tract_xwalk(
  data_in_name = NULL,
  data_in_pth = NULL,
  pth = "https://www.datahouston.org/cta_crosswalk.txt"
)
}
\arguments{
\item{data_in_name}{string. The name to of the data to read in.}

\item{data_in_pth}{string. The path to read the data in from.}
}
\description{
Download the neighbordhood to census tract cross walk create by the Kinder Institute. https://www.arcgis.com/apps/MapSeries/index.html?appid=95320b06677c438d91027cb5feb241bf
}
