% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get.nber_tract_data}
\alias{get.nber_tract_data}
\title{Get NBER Tract data}
\usage{
get.nber_tract_data(
  data_in_name,
  data_in_pth,
  url = "http://data.nber.org/distance/2010/sf1/tract/{fl}"
)
}
\arguments{
\item{data_in_name}{string. The name of the file to download.}

\item{data_in_pth}{string. The path to read the data in from.}
}
\description{
Link to data: http://data.nber.org/distance/2010/sf1/tract/sf12010tractdistance25miles.csv
}
\details{
25 mile radius is used rather then the 5 miles radius because the 5 mile radius was missing two counties in Harris County
}
