% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get.state_fips_state_name_xwalk}
\alias{get.state_fips_state_name_xwalk}
\title{Get State FIPS and State Name Crosswalk}
\usage{
get.state_fips_state_name_xwalk(data_in_name, data_in_pth, state_fips = 48)
}
\arguments{
\item{data_in_name}{string. The name to of the data to read in.}

\item{data_in_pth}{string. The path to read the data in from.}
}
\description{
Downloads the crosswalk between county fips codes and county names for Texas
}
