% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get.tract_shape}
\alias{get.tract_shape}
\title{Get tract by latitude and longitude}
\usage{
get.tract_shape(data_in_name, data_in_pth, state_fips = 48, county_fips = 201)
}
\arguments{
\item{data_in_name}{string. The name to of the data to read in.}

\item{data_in_pth}{string. The path to read the data in from.}
}
\description{
Downloads shape file for Harris County (201) Texas (48) using the tigris package, which pulls the most recent shape from the United States Census Bureau.
}
