% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get.zip_county_xwalk}
\alias{get.zip_county_xwalk}
\title{Get county and ZCTA (zip code) Crosswalk}
\usage{
get.zip_county_xwalk(
  data_in_name = NULL,
  data_in_pth = NULL,
  pth = "https://www2.census.gov/geo/docs/maps-data/data/rel/zcta_county_rel_10.txt",
  state_fips = 48
)
}
\arguments{
\item{data_in_name}{string. The name to of the data to read in.}

\item{data_in_pth}{string. The path to read the data in from.}

\item{pth}{Link to data: https://www2.census.gov/geo/docs/maps-data/data/rel/zcta_county_rel_10.txt}
}
\description{
Downloads the crosswalk between county fips codes and county names for Texas.
}
