% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get.zip_latlong_xwalk}
\alias{get.zip_latlong_xwalk}
\title{Get ZCTA (Zip Code) and Latitude/Longitude Coordinate Crosswalk}
\usage{
get.zip_latlong_xwalk(data_in_pth = NULL, data_in_name = NULL, state_fips = 48)
}
\arguments{
\item{data_in_pth}{string. The path to read the data in from.}

\item{data_in_name}{string. The name to of the data to read in.}
}
\description{
Downloads the crosswalk between zip codes and latitude longitude coordinates of the center for Texas
}
