% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_landmarks.R
\name{get_geo.highway}
\alias{get_geo.highway}
\title{Get Geo highways}
\usage{
get_geo.highway(
  county_name,
  key = "highway",
  value = "motorway",
  geo_type = "osm_lines"
)
}
\description{
Get Geo highways
}
