% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_landmarks.R
\name{get_geo.waterway}
\alias{get_geo.waterway}
\title{Get Geo waterways}
\usage{
get_geo.waterway(
  county_name,
  key = "waterway",
  value = "river",
  geo_type = "osm_lines"
)
}
\description{
Get Geo waterways
}
