% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helper_functions.R
\name{get_key.geocodio}
\alias{get_key.geocodio}
\title{Checks to make sure the geocodio API key is set as an environment 
variable}
\usage{
\method{get_key}{geocodio}(
  env_var = "Geocodio_API_KEY",
  title = "Geocodio",
  url = "https://dash.geocod.io/apikey/create",
  install = "texascc::api_key.geocodio(key = 'X')"
)
}
\arguments{
\item{env_var.}{String. The environment variable to check for.}
}
\value{
TRUE if a census api key exists, error if it does not exist
}
\description{
Checks to make sure the geocodio API key is set as an environment 
variable
}
\examples{
\dontrun{
texascc::api_key.geocodio(key = "X")
texascc::get_key.geocodio()
}
}
