% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helper_functions.R
\name{get_key.google}
\alias{get_key.google}
\title{Checks to make sure the census API key is set as an environment variable}
\usage{
\method{get_key}{google}(
  env_var = "Google_API_KEY",
  title = "Goggle",
  url = "https://developer.mapquest.com/",
  install = "texascc::api_key.google(key = 'X')"
)
}
\arguments{
\item{env_var.}{String. The environment variable to check for.}
}
\value{
TRUE if a census api key exists, error if it does not exist
}
\description{
Checks to make sure the census API key is set as an environment variable
}
\examples{
\dontrun{
texascc::api_key.google(key = "X")
texascc::get_key.google()
}
}
