% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkt_analysis.R
\name{high_need_table}
\alias{high_need_table}
\title{Find high need table for each county}
\usage{
high_need_table(county_fips, neighborhood_attrs, zip_attrs)
}
\description{
Find high need table for each county
}
