% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{neighborhood_attributes}
\alias{neighborhood_attributes}
\title{Find demand for each neighborhood}
\usage{
neighborhood_attributes(
  neighborhood_desert,
  neighborhood_demand,
  neighborhood_svi
)
}
\description{
Find demand for each neighborhood
}
