% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{neighborhood_demand}
\alias{neighborhood_demand}
\title{Find demand for each neighborhood}
\usage{
neighborhood_demand(xwalk_neighborhood_tract, tract_demand)
}
\description{
Find demand for each neighborhood
}
