% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkt_analysis.R
\name{neighborhood_desert}
\alias{neighborhood_desert}
\title{Find neighborhoods which are deserts}
\usage{
neighborhood_desert(xwalk_neighborhood_tract, df_ratio)
}
\description{
Find neighborhoods which are deserts
}
