% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{neighborhood_svi}
\alias{neighborhood_svi}
\title{Find SVI stats for each neighborhood}
\usage{
neighborhood_svi(xwalk_neighborhood_tract, tract_svi)
}
\description{
Find SVI stats for each neighborhood
}
