% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{parse_date.frontline}
\alias{parse_date.frontline}
\title{Parse frontline date}
\usage{
parse_date.frontline(x, date_format = "\\\\d{4}-\\\\d{2}-\\\\d{2}")
}
\arguments{
\item{format}{string. The format to use to parse the string. Default is
'%m.%d.%y'.}
}
\value{
date
}
\description{
Parse frontline date
}
