% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{parse_date.twc}
\alias{parse_date.twc}
\title{Parse TWC date}
\usage{
parse_date.twc(x, date_format = "\%m.\%d.\%y")
}
\arguments{
\item{format}{string. The format to use to parse the string. Default is
'%m.%d.%y'.}
}
\value{
date
}
\description{
Parse TWC date
}
