% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkt_analysis.R
\name{pct_desert}
\alias{pct_desert}
\title{Find percent of markets that are a desert for a specific type of desert}
\usage{
pct_desert(df)
}
\description{
Find percent of markets that are a desert for a specific type of desert
}
