% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{pop.hhsc_ccl}
\alias{pop.hhsc_ccl}
\title{HHSC CCL Population}
\usage{
pop.hhsc_ccl(new, old)
}
\value{
data.frame
}
\description{
Creates a dataframe unique on operation number and download date
to be able to track when child care providers enter and leave the CCL database
}
