% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{pop.hhsc_ccl_most_recent_attr}
\alias{pop.hhsc_ccl_most_recent_attr}
\title{HHSC CCL Population}
\usage{
pop.hhsc_ccl_most_recent_attr(new, old)
}
\value{
data.frame
}
\description{
Keeps to most recent attributes for each provider by download
date
}
