% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts_dm.R
\name{process.adj_tracts}
\alias{process.adj_tracts}
\title{Process adjacent tracts}
\usage{
process.adj_tracts(cls, fl = "nlist_2010.csv")
}
\description{
Data from: https://s4.ad.brown.edu/Projects/Diversity/Researcher/Pooling.htm
}
