% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarrant_neighborhood.R
\name{process.harris_neighborhood}
\alias{process.harris_neighborhood}
\title{Process the Tarrant Neighborhood data}
\usage{
process.harris_neighborhood(cls)
}
\description{
Process the Tarrant Neighborhood data
}
