% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{process.tarrant_neighborhood}
\alias{process.tarrant_neighborhood}
\title{Process the Tarrant Neighborhood data}
\usage{
process.tarrant_neighborhood(cls)
}
\description{
Process the Tarrant Neighborhood data
}
