% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{process.xwalk_neighborhood_tract}
\alias{process.xwalk_neighborhood_tract}
\title{Process Neighborhood Tract Xwalk}
\usage{
process.xwalk_neighborhood_tract(raw_pth)
}
\description{
Process Neighborhood Tract Xwalk
}
\note{
Harris County only because of data availability
}
