% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{process.xwalk_tract_prvdr}
\alias{process.xwalk_tract_prvdr}
\title{Process the Tracts and Provider crosswalk}
\usage{
process.xwalk_tract_prvdr(xwalk_tracts, df_hhsc_ccl)
}
\description{
Process the Tracts and Provider crosswalk
}
