% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkt_analysis.R
\name{prvdrs_serving_high_need}
\alias{prvdrs_serving_high_need}
\title{Find providers serving high need areas}
\usage{
prvdrs_serving_high_need(
  county_fips,
  high_need,
  xwalk_neighborhood_tract,
  xwalk_zip_tract,
  df_supply,
  providers
)
}
\description{
Find providers serving high need areas
}
