% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_file_from_dwnld}
\alias{read_file_from_dwnld}
\title{Read file from download}
\usage{
read_file_from_dwnld(temp, unzip_fl, pth, df_name, ...)
}
\arguments{
\item{temp.}{Object. The temporary file which is downloaded}

\item{unzip_fl.}{logical. Unzips a file if the file is zipped.}
}
\value{
data.frame
}
\description{
Read file from download
}
