% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_fl_ext}
\alias{read_fl_ext}
\title{Read file by extension}
\usage{
read_fl_ext(pth, ...)
}
\arguments{
\item{sheet.}{integer. The sheet to read in. If it's a csv then sheet is set 
to NULL}

\item{skip.}{integer. The number of rows to skip when reading in}
}
\value{
dataframe
}
\description{
Read file by extension
}
