% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_child_care_db.R
\name{save_subset_child_care_db}
\alias{save_subset_child_care_db}
\title{Save a subset of the child care database}
\usage{
save_subset_child_care_db(pth, config)
}
\arguments{
\item{pth}{string. Path to the root directory to create the DB.}

\item{county}{string. County 5-digit FIPS code.}
}
\description{
Save a subset of the child care database
}
\examples{
\dontrun{
pth <- "C:/"
county <- "48439"
save_subset_child_care_db(pth = pth, county = county)
}
}
