% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_child_care_db.R
\name{save_subset_child_care_db_03_group}
\alias{save_subset_child_care_db_03_group}
\title{Save a subset of the child care database for children 0-3 for a group of counties}
\usage{
save_subset_child_care_db_03_group(pth, config, dev = TRUE)
}
\arguments{
\item{pth}{string. Path to the root directory to create the DB.}

\item{config}{object. Object containing parameters to create the DB. 
Config should be named list with 3 components: "label" for naming, "county_names" list of counties 
to include, "tract_radius"}

\item{dev}{boolean. St dev = FALSE to save the db to the production environment. Default is TRUE.}
}
\description{
Save a subset of the child care database for children 0-3 for a group of counties
}
