% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_dm.R
\name{select_qtr_year}
\alias{select_qtr_year}
\title{Select quarter year}
\usage{
select_qtr_year(pth, acf_qtr_years)
}
\arguments{
\item{pth}{String. Path to the data}

\item{acf_qtr_years}{vector. Default is 'Q2-2019'.}
}
\value{
A vector of file paths associated with the given quarter year
}
\description{
Parses the quarter-year parameter and returns the file names
which are associated with that quarter year combination
}
