% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_dm.R
\name{split_calls}
\alias{split_calls}
\title{Split calls}
\usage{
split_calls(v, limit)
}
\arguments{
\item{v}{vector. Vector to split into multiple groups}

\item{limit}{integer. The max size of each group.}
}
\description{
Splits list into multiple groups for batch calls with limits per call
}
