% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subsidy_capacity.R
\name{subset_tracts}
\alias{subset_tracts}
\title{Subset the tract crosswalk crosswalk}
\usage{
subset_tracts(xwalk_tracts, adj_tracts, tract_radius, county_fips)
}
\description{
Subset the tract crosswalk crosswalk
}
