% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tract_radius.R
\name{summary.acf_dist}
\alias{summary.acf_dist}
\title{ACF data management to get zip codes for providers}
\usage{
\method{summary}{acf_dist}(df, county_list)
}
\arguments{
\item{df}{data.frame.}

\item{county_list.}{A list of TX county codes of counties of interest to 
compare. E.g county_list = c(48439) for Tarrant County; 
county_list = c("48201", "48439") to compare Harris to Tarrant County}
}
\description{
Clean ACF provider data, find coordinates for each zip code, 
calculate straight line distance.
}
