% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_joins.R
\name{surround_tract_sub}
\alias{surround_tract_sub}
\title{Filter to corresponding surround tracts to the anchor tract}
\usage{
surround_tract_sub(df_anchor_tract, xwalk_tracts)
}
\arguments{
\item{xwalk_tracts}{data.frame. Contains the filtered anchor tracts.}

\item{df_tracts_xwalks}{data.frame. Crosswalk between the anchor and surround}

\item{input_mkt}{data.frame. Data frame with subset of tracts to map}
}
\description{
Filter to corresponding surround tracts to the anchor tract
}
