% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{test.hhsc_ccl}
\alias{test.hhsc_ccl}
\title{Test ccl data management function}
\usage{
test.hhsc_ccl(df)
}
\arguments{
\item{df}{data.frame. A data frame downloaded from get.hhsc_ccl_data}
}
\description{
Tests that data management function works correctly
}
