% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subsidy_capacity.R
\name{test_config}
\alias{test_config}
\title{Test config}
\usage{
test_config(
  x,
  str,
  msg = "Parameter '{str}' is missing from {n} county in the list"
)
}
\description{
Test to make sure configuration is set-up correctly
}
