% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{test_config_pct}
\alias{test_config_pct}
\title{Test config pct}
\usage{
test_config_pct(x, str, msg = "Parameter '{str}' should be between 0 and 1")
}
\description{
Test percent parameters are between 0 and 1
}
