% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cols.R
\name{test_input}
\alias{test_input}
\title{Test and select input}
\usage{
test_input(df, input_columns)
}
\arguments{
\item{df}{data.frame}

\item{input_columns.}{List. List of variable names to select with expected
variable types.}
}
\value{
data.frame
}
\description{
Tests that the expected columns exist, and then selects those
columns
}
