% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cols.R
\name{test_type.character}
\alias{test_type.character}
\title{Test for character (string) column or vector}
\usage{
\method{test_type}{character}(v)
}
\description{
Test for character (string) column or vector
}
