% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkt_analysis.R
\name{total_children}
\alias{total_children}
\title{Find total number of children in need of care of each type}
\usage{
total_children(df)
}
\description{
Find total number of children in need of care of each type
}
