% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkt_analysis.R
\name{total_children_desert}
\alias{total_children_desert}
\title{Find number of children per desert type}
\usage{
total_children_desert(df_ratio, df_demand)
}
\description{
Find number of children per desert type
}
