% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{verify_200_status}
\alias{verify_200_status}
\title{Verify 200 status.}
\usage{
verify_200_status(status, temp, pth, df_name, unzip_fl, ...)
}
\arguments{
\item{status}{integer. The status of the request.}

\item{pth.}{string. The path to write the data to.}

\item{df_name.}{string. The name of the file to write the data to.}

\item{unzip.}{logical. TRUE if the file needs to be unzipped.}
}
\description{
Verify that the status of the request is 200.
}
