% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subsidy_capacity.R
\name{calc.subsidy_capacity}
\alias{calc.subsidy_capacity}
\title{Create subsidy capacity estimate}
\usage{
calc.subsidy_capacity(
  county,
  tract_radius,
  xwalk_tracts,
  adj_tracts,
  df_hhsc_ccl,
  df_acf,
  qtrs
)
}
\description{
Create subsidy capacity estimate
}
