% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_child_care_db.R
\name{childcare_db}
\alias{childcare_db}
\title{Create child care data base}
\usage{
childcare_db(
  root,
  acs_year = 2019,
  acs_state_code = 48,
  acs_geography = "tract",
  acs_county = 439
)
}
\arguments{
\item{root}{string. Path to the root directory to create the DB}

\item{acs_year.}{Integer. ACS year to pull the tables from. Default is 2019.}

\item{acs_state_code.}{Integer. ACS state_code to pull the data for. Default 
is 48 (Texas).}

\item{acs_geography.}{String. ACS geography. Default is "tract".}

\item{acs_county.}{Integer. County FIPS code (3-digits). Default is 439 
(Tarrant County).}
root <- "C:/"
childcare_db(root = root)
}
