% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_child_care_db.R
\name{childcare_db}
\alias{childcare_db}
\title{Create child care data base}
\usage{
childcare_db(acs_tbls, root)
}
\arguments{
\item{root}{string. Path to the root directory to create the DB}

\item{acs_tbls.}{List of census tables to pull}
}
\description{
Create child care data base
}
\examples{
\dontrun{
census_tbls <- list(B23008 = list(year = 2019, state = 48, 
                    geography = "tract", county = 439))
root <- "C:/"
childcare_db(acs_tbls = acs_tbls, root = root)
}
}
