% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_child_care_db.R
\name{childcare_db}
\alias{childcare_db}
\title{Create child care data base}
\usage{
childcare_db(root, acf_qtr_years = "Q2-2019")
}
\arguments{
\item{root}{string. Path to the root directory to create the DB.}

\item{acf_qtr_years}{vector. Default is 'Q2-2019'.}
}
\description{
Create child care data base
}
\examples{
\dontrun{
root <- "C:/"
childcare_db(root = root)
}
}
