% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.accepts_child_care_subsidies}
\alias{col.accepts_child_care_subsidies}
\title{Data management steps to clean the accepts child care subsidies column}
\usage{
col.accepts_child_care_subsidies(df)
}
\arguments{
\item{df}{data.frame. The dataframe}
}
\value{
data.frame
}
\description{
Data management steps to clean the accepts child care subsidies column
}
