% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.county}
\alias{col.county}
\title{Data management steps for the county column}
\usage{
col.county(df)
}
\arguments{
\item{df}{data.frame. The dataframe}
}
\value{
data.frame
}
\description{
Data management steps for the county column
}
