% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.location_address_geo}
\alias{col.location_address_geo}
\title{Data management steps to clean the operation type column}
\usage{
col.location_address_geo(df, state_fips)
}
\arguments{
\item{df}{data.frame. The dataframe}
}
\value{
data.frame
}
\description{
Data management steps to clean the operation type column
}
