% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{col.mod_date}
\alias{col.mod_date}
\title{Data management for modification date column}
\usage{
col.mod_date(x, date_format = "\%m\%d\%Y")
}
\value{
object
}
\description{
Returns a data.frame with the most recent modification date for 
each provider.
}
