% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.operation_name}
\alias{col.operation_name}
\title{Data management steps to clean the operation name column}
\usage{
col.operation_name(df)
}
\arguments{
\item{df}{data.frame. The dataframe}
}
\value{
data.frame
}
\description{
Data management steps to clean the operation name column
}
