% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{col.operation_type}
\alias{col.operation_type}
\title{Data management steps to clean the location address geo column}
\usage{
col.operation_type(df)
}
\arguments{
\item{df}{data.frame. The dataframe}
}
\value{
data.frame
}
\description{
Data management steps to clean the location address geo column
}
