% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tract_radius.R
\name{distance_between_zips}
\alias{distance_between_zips}
\title{Calculate distance between two zip codes in miles}
\usage{
distance_between_zips(df, zip1, zip2)
}
\arguments{
\item{df.}{The data frame containing matched zip code columns.}

\item{zip1.}{One of the zip code columns.}

\item{zip2.}{The other zip code column.}
}
\value{
data.frame with distance column.
}
\description{
Find lat/lon coordinates for each zip code, calculate straight line distance.
}
