% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_dm.R
\name{dm.agg_kids_prvdr}
\alias{dm.agg_kids_prvdr}
\title{Aggregate number of kids per provider}
\usage{
dm.agg_kids_prvdr(acf)
}
\arguments{
\item{acf}{data.frame. The cleaned acf dataframe.}
}
\value{
Summarized data with the max, median, and minimum number of kids per
provider
}
\description{
Aggregate number of kids per provider
}
