% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subsidy_capacity.R
\name{dm.agg_ratio_mkt}
\alias{dm.agg_ratio_mkt}
\title{Calculate enrollment ratios aggregated to market level}
\usage{
\method{dm}{agg_ratio_mkt}(n_kids, df_hhsc_ccl, xwalk_tract_prvdr)
}
\arguments{
\item{n_kids}{data.frame. Result of dm.agg_kids_prvdr.}
}
\description{
Calculate enrollment ratios aggregated to market level
}
