% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_dm.R
\name{dm.demand}
\alias{dm.demand}
\title{Data management steps to create the demand table}
\usage{
\method{dm}{demand}(B23008, B17024, processed_pth, name = "demand", pov_rate = 0.85)
}
\arguments{
\item{pov_rate}{children w/ working parents are (1-pov_rate) more likely 
to be under 200% of the poverty line. Default is .85.}
}
\description{
Children in poverty with working parents
}
