% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{dm.frontline}
\alias{dm.frontline}
\title{Data management for frontline provider data}
\usage{
\method{dm}{frontline}(
  df,
  input_columns = list(OP_Number = "character", Infant_Capacity = "numeric",
    Toddler_Capacity = "numeric", Pre_K_Capacity = "numeric", School_Aged_Capacity =
    "numeric", Infant_enrollment = "numeric", Toddler_enrollment = "numeric",
    Preschool_enrollment = "numeric", SchoolAge_enrollment = "numeric",
    last_modified_at_A = "character", Date = "character", export_date = "Date",
    Reporting_Status = "character")
)
}
\description{
Manage frontline data, specifically enrollment and attendance numbers
}
