% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{dm.hhsc_ccl}
\alias{dm.hhsc_ccl}
\title{HHSC CCL data management}
\usage{
dm.hhsc_ccl(df, input_columns, download_date = Sys.Date(), ...)
}
\arguments{
\item{county_name}{string. The name of the county to subset to. Default is NULL.}
}
\value{
data.frame
}
\description{
Clean CCL download data, convert key variables to binary and select
}
