% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_dm.R
\name{dm.reverse_geocode}
\alias{dm.reverse_geocode}
\title{Geocode addresses}
\usage{
\method{dm}{reverse_geocode}(df, url = "https://geo.fcc.gov", path = "/api/census/block/find")
}
\description{
Geocodes addresses using FCC
}
