% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{dm.twc}
\alias{dm.twc}
\title{Data management for TWC provider data}
\usage{
\method{dm}{twc}(x)
}
\description{
Manage TWC data, specifically TRS status and level, subsidy status, and
number of subsidy kids served
}
