% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_dm.R
\name{dwnld.acs}
\alias{dwnld.acs}
\title{Download ACS}
\usage{
dwnld.acs(tbls, pth)
}
\arguments{
\item{tbls}{list. List of census tables with attributes to download}

\item{pth}{string. The path to download the data to.}
}
\description{
Passes in a list of parameters to download the ACS census
data using functions from the tidycensus package
}
\details{
To find a list of parameters to pass see documentation for 
tidycensus::get_acs
}
\examples{
\dontrun{
census_tbls <- list(B23008 = list(year = 2019, state = 48, 
                                  geography = "tract", county = "48438"))
pth <- "C:/"
dwnld.acs(tbls = census_tbls, pth = pth)
}
}
