% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{dwnld.frontline}
\alias{dwnld.frontline}
\title{Download Frontline provider data}
\usage{
dwnld.frontline(
  raw_pth,
  name = "frontline/export_translation_Daily_Vacancy_2021-08-30_04_43_51.csv"
)
}
\description{
Frontline provider data comes from the childcare.bowtiebi.com portal (currently, only Sadie has login)
}
