% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontline_dm.R
\name{dwnld.frontline}
\alias{dwnld.frontline}
\title{Download Frontline provider data}
\usage{
dwnld.frontline(
  raw_pth,
 
    name = "frontline/Rice File Given on 11.16.21_Daily_Vacancy_Golden_Sheet_2021-11-16.xlsx",
  sheet = "Rice File"
)
}
\description{
Frontline provider data comes from the childcare.bowtiebi.com portal. Current sheet was emailed by Myriam Guillen 11/16.
}
