% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwnld.hhsc_ccl}
\alias{dwnld.hhsc_ccl}
\title{Get HHSC CCL data}
\usage{
dwnld.hhsc_ccl(
  name,
  raw_pth,
  url = "https://data.texas.gov/api/views/bc5r-88dy/rows.csv?accessType=DOWNLOAD",
  ext = "csv",
  ...
)
}
\arguments{
\item{name}{string. The name to write the raw data to.}

\item{raw_pth}{string. The path to write the raw to.}

\item{url}{string. The url to the data.}
}
\description{
Returns the most recent HHSC CCL Daycare and Residential 
Operations 
Data. Link to data: https://data.texas.gov/resource/bc5r-88dy.csv
}
