% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccl_dm.R
\name{dwnld.hhsc_ccl}
\alias{dwnld.hhsc_ccl}
\title{Get HHSC CCL data}
\usage{
dwnld.hhsc_ccl(x)
}
\description{
Returns the most recent HHSC CCL Daycare and Residential 
Operations 
Data. Link to data: https://data.texas.gov/resource/bc5r-88dy.csv
}
