% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc_dm.R
\name{dwnld.twc}
\alias{dwnld.twc}
\title{Download TWC provider data}
\usage{
dwnld.twc(
  raw_pth,
  name = "Closure Report 11.8.21.xlsx",
  sheet = "All Provider Level"
)
}
\description{
TWC provider data is emailed by Shay when we request
}
