% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helper_functions.R
\name{get_key.census}
\alias{get_key.census}
\title{Checks to make sure the census API key is set as an environment variable}
\usage{
\method{get_key}{census}(
  env_var = "CENSUS_API_KEY",
  title = "Census",
  url = "https://api.census.gov/data/key_signup.html",
  install = "tidycensus::census_api_key(key = 'X')"
)
}
\value{
TRUE if a census api key exists, error if it does not exist
}
\description{
Checks to make sure the census API key is set as an environment variable
}
\examples{
\dontrun{
tidycensus::census_api_key(key = "X")
texascc::get_key.census()
}
}
