% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_child_care_db.R
\name{save_subset_child_care_db_03}
\alias{save_subset_child_care_db_03}
\title{Save a subset of the child care database for children 0-3}
\usage{
save_subset_child_care_db_03(pth, config, dev = TRUE)
}
\arguments{
\item{pth}{string. Path to the root directory to create the DB.}

\item{config}{object. Object containing parameters to create the DB.}

\item{dev}{boolean. St dev = FALSE to save the db to the production environment. Default is TRUE.}
}
\description{
Save a subset of the child care database for children 0-3
}
\examples{
\dontrun{
pth <- "C:/"
county <- "48439"
save_subset_child_care_db_03(pth = pth, county = county)
}
}
