% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tract_radius.R
\name{summary.acf_dist}
\alias{summary.acf_dist}
\title{ACF data management to get zip codes for providers}
\usage{
\method{summary}{acf_dist}(df, county_list = NULL)
}
\arguments{
\item{df}{data.frame. The merged family/home ACF data from dm.acf_dist 
function.}

\item{county_list.}{A list of TX county codes of counties of interest to 
compare. 
Eg county_list = c(439) for Tarrant County; county_list = c(201, 439) to 
compare Harris to Tarrant County}
}
\value{
a table and a plot. Decile values for distance between home and 
provider zip code and density plot of distances.
}
\description{
Clean ACF provider data, find coordinates for each zip code, 
calculate straight line distance.
}
